feat(betterportal): scale teleport locations
In the vanilla game, 8 blocks in the nether is equal to 1 block in the overworld. This requires the nether world border to be 8 times bigger than that of overworld, so that the nether portals can work as intended. But, we don't want that big nether world border in the server as that would take too much disk space.

We used to simply cancel the event when portal's to location is outside the world border. This caused some issues where either player would just be stuck in the nether world, or they cannot go to the nether from the overworld.

To ultimately solve this issue, we scale the portal's to location to make sure the to location is within the border.